Move TestOptions methods to Test
